Extract CommonBaseCard
