+++ src/components/elements/CommonCollapsibleCard.stories.tsx
@@ -1,11 +1,7 @@
-import { Button } from '@mui/material';
 import { useArgs } from '@storybook/preview-api';
 import { Meta, StoryObj } from '@storybook/react';
 
 import CommonCollapsibleCard from './CommonCollapsibleCard';
-import GenericTable from '@/components/elements/table/GenericTable';
-import { DummyTableRows } from '@/components/elements/table/GenericTable.stories';
-import { CLIENT_COLUMNS } from '@/modules/search/components/ClientSearch';
 
 export default {
   component: CommonCollapsibleCard,
@@ -39,31 +35,30 @@ const cardContent = (
   </>
 );
 
+// export const Default: Story = {
+//   args: {
+//     title: 'Card with Action Button',
+//     children: cardContent,
+//     actions: <Button>Click Me</Button>,
+//   },
+// };
+
+// export const TableContent: Story = {
+//   args: {
+//     title: 'Card with Table Content (No Padding)',
+//     padContent: false,
+//     children: (
+//       <GenericTable
+//         rows={DummyTableRows}
+//         columns={[CLIENT_COLUMNS.name, CLIENT_COLUMNS.age]}
+//       />
+//     ),
+//   },
+// };
+
 export const Default: Story = {
   args: {
-    title: 'Card with Action Button',
-    children: cardContent,
-    actions: <Button>Click Me</Button>,
-  },
-};
-
-export const TableContent: Story = {
-  args: {
-    title: 'Card with Table Content (No Padding)',
-    padContent: false,
-    children: (
-      <GenericTable
-        rows={DummyTableRows}
-        columns={[CLIENT_COLUMNS.name, CLIENT_COLUMNS.age]}
-      />
-    ),
-  },
-};
-
-export const Collapsible: Story = {
-  args: {
     title: 'Card that is Collapsible',
-    collapsible: true,
     children: cardContent,
   },
 };
